+++ amd/src/local/content/actions.js
@@ -384,7 +384,7 @@ export default class extends BaseComponent {
     }
 
     /**
-     * Handle a move request to add a subsection.
+     * Handle a request to add a subsection as the last child of the parent
      *
      * @param {Element} target the dispatch action element
      * @param {Event} event the triggered event
@@ -392,6 +392,17 @@ export default class extends BaseComponent {
     _requestAddSubSection(target, event) {
         event.preventDefault();
         this.reactive.dispatch('addSubSection', parseInt(target.dataset.parentid ?? 0));
+    }
+
+    /**
+     * Handle a request to add a subsection as the first child of the parent
+     *
+     * @param {Element} target the dispatch action element
+     * @param {Event} event the triggered event
+     */
+    _requestInsertSubSection(target, event) {
+        event.preventDefault();
+        this.reactive.dispatch('insertSubSection', parseInt(target.dataset.parentid ?? 0));
     }
 
     /**